preparing for adding plan of study
