+++ db-service/db-service.iml
@@ -15,7 +15,7 @@
         <setting name="validation-enabled" value="true" />
         <setting name="provider-name" value="Hibernate" />
         <datasource-mapping>
-          <factory-entry name="entityManagerFactory" />
+          <factory-entry name="entityManagerFactory" value="a8c8e562-ae19-4b37-b7de-b062be819ff7" />
         </datasource-mapping>
         <naming-strategy-map />
       </configuration>
