+++ db-service/src/main/java/com/flux/dbservice/entity/parsing/lessons/Audience.java
@@ -1,4 +1,4 @@
-package com.flux.dbservice.entity.parsing;
+package com.flux.dbservice.entity.parsing.lessons;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -13,8 +13,8 @@ import javax.persistence.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
-@Table(name = "teacher", schema = "parsing")
-public class Teacher {
+@Table(name = "audience", schema = "lesson_parsing")
+public class Audience {
 
     @Id
     @GeneratedValue(strategy = GenerationType.TABLE)
