+++ db-service/src/main/java/com/flux/dbservice/repository/parsing/AudienceRepository.java
@@ -1,6 +1,6 @@
 package com.flux.dbservice.repository.parsing;
 
-import com.flux.dbservice.entity.parsing.Audience;
+import com.flux.dbservice.entity.parsing.lessons.Audience;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
