+++ db-service/src/main/java/com/flux/dbservice/repository/parsing/GroupRepository.java
@@ -1,6 +1,6 @@
 package com.flux.dbservice.repository.parsing;
 
-import com.flux.dbservice.entity.parsing.Group;
+import com.flux.dbservice.entity.parsing.lessons.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
