+++ telegram-user-service/src/main/java/com/flux/telegramservice/service/request/RestTemplateService.java
@@ -1,6 +1,7 @@
 package com.flux.telegramservice.service.request;
 
 import com.flux.telegramservice.entity.*;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
